+++ CHANGELOG.md
@@ -5,8 +5,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add rsync in Server scripts
+- Debian Family Desktop srcipts
+- add tmux bashcompletion
+- add timetrap in Desktop scripts 
+- add rbenv in Desktop scripts
+- add htop in the list of additional software to install
+- add ncdu in the list of additional software to install
+- add EPEL repo to Rocky linux post-install
+- 
+### Changed
+- disable cockpit in desktop post-install
 
 ## [0.1.0] - 2021-10-16
 ### Added
-- Rocky Linux Desktop and Server post-install script.
+- Rocky Linux Desktop and Server post-install script
 - Raspberry Pi OS post-install script